refactor(seed): simplify BYOK evidence helpers
