+++ dev/seed/lib/users.ts
@@ -1,0 +1,41 @@
+import { kilocode_users } from '@kilocode/db/schema';
+import { eq, or } from 'drizzle-orm';
+
+import { getSeedDb } from './db';
+import { normalizeSeedEmail } from './email';
+
+export function isValidEmail(email: string): boolean {
+  // Intentionally permissive; we only guard against obvious nonsense in dev.
+  return /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(email);
+}
+
+export async function resolveSeedUserId(email: string): Promise<string> {
+  const normalizedEmail = normalizeSeedEmail(email);
+  const db = getSeedDb();
+  const matches = await db
+    .select({
+      userId: kilocode_users.id,
+      email: kilocode_users.google_user_email,
+    })
+    .from(kilocode_users)
+    .where(
+      or(
+        eq(kilocode_users.google_user_email, email),
+        eq(kilocode_users.normalized_email, normalizedEmail)
+      )
+    );
+
+  if (matches.length === 0) {
+    throw new Error(`No user found for email ${email}`);
+  }
+
+  const exactMatches = matches.filter(match => match.email === email);
+  const resolvedMatches = exactMatches.length > 0 ? exactMatches : matches;
+  if (resolvedMatches.length > 1) {
+    const matchList = resolvedMatches.map(match => `${match.email} (${match.userId})`).join(', ');
+    throw new Error(`Multiple users matched ${email}: ${matchList}`);
+  }
+
+  const [user] = resolvedMatches;
+  return user.userId;
+}
